+++ packages/cloud-cognitive/src/components/ImportModal/ImportModal.stories.js
@@ -1,5 +1,5 @@
 //
-// Copyright IBM Corp. 2020, 2020
+// Copyright IBM Corp. 2021, 2021
 //
 // This source code is licensed under the Apache-2.0 license found in the
 // LICENSE file in the root directory of this source tree.
@@ -7,6 +7,7 @@
 
 import React, { useState } from 'react';
 import { Button } from 'carbon-components-react';
+import { action } from '@storybook/addon-actions';
 import styles from './_storybook-styles.scss'; // import index in case more files are added later.
 import { pkg } from '../../settings';
 import { getStorybookPrefix } from '../../../config';
@@ -26,6 +27,8 @@ export default {
 };
 
 const defaultProps = {
+  body: 'You can specify a file to import by either dragging it into the drag and drop area or by specifying a URL. (Maximum file size of 500KB; .jpg and .png file extensions only.)',
+  className: 'test-class',
   defaultErrorBody: 'Select a new file and try again.',
   defaultErrorHeader: 'Import failed',
   fetchErrorBody: 'Unable to fetch URL.',
@@ -34,50 +37,35 @@ const defaultProps = {
   fileDropLabel: 'Drag and drop files here or click to upload',
   fileUploadLabel: 'files uploaded',
   inputButtonText: 'Add file',
-  inputHeader: 'Add a file by specifying a URL',
   inputId: 'test-id',
+  inputLabel: 'Add a file by specifying a URL',
   inputPlaceholder: 'URL',
   invalidFileTypeErrorBody: 'Invalid file type.',
   invalidFileTypeErrorHeader: 'Import failed',
+  invalidIconDescription: 'Delete',
   maxFileSize: 500000,
   maxFileSizeErrorBody: '500kb max file size. Select a new file and try again.',
   maxFileSizeErrorHeader: 'Import failed',
-  modalBody:
-    'You can specify a file to import by either dragging it into the drag and drop area or by specifying a URL. (Maximum file size of 500KB; .jpg and .png file extensions only.)',
-  modalHeading: 'Import',
-  multiple: false,
-  onRequestClose: () => console.log('closed'),
+  onClose: action('onClose event'),
   onRequestSubmit: (file) => console.log('file contents', file),
   open: true,
   primaryButtonText: 'Import',
   secondaryButtonText: 'Cancel',
+  title: 'Import',
   validFileTypes: ['image/jpeg', 'image/png'],
-};
-
-const Template = (args) => {
-  return <ImportModal {...args} />;
 };
 
 const TemplateWithState = (args) => {
   const [open, setOpen] = useState(false);
   return (
     <>
-      <ImportModal
-        {...args}
-        open={open}
-        onRequestClose={() => setOpen(false)}
-      />
+      <ImportModal {...args} open={open} onClose={() => setOpen(false)} />
       <Button onClick={() => setOpen(true)}>Launch modal</Button>
     </>
   );
 };
 
-export const WithStateManager = TemplateWithState.bind({});
-WithStateManager.args = {
-  ...defaultProps,
-};
-
-export const Standard = Template.bind({});
+export const Standard = TemplateWithState.bind({});
 Standard.args = {
   ...defaultProps,
 };